requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
aiomysql==0.0.20
altgraph==0.16.1
appdirs==1.4.4
argcomplete==1.0.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
bar==0.2.1
beautifulsoup4==4.8.2
bleach==3.1.0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
colorama==0.4.1
cryptography==2.9.2
cx-Freeze==6.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.8
django-allauth==0.42.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-taggit==1.3.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.7.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
multidict==4.7.6
multitasking==0.0.9
mysql-connector==2.2.9
mysqlclient==2.0.1
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.2.0.34
openpyxl==3.0.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.3.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
pycparser==2.20
pygame==1.9.6
Pygments==2.4.2
PyInstaller==3.5
PyMySQL==0.10.0
pyparsing==2.4.5
pyrsistent==0.15.3
PySDL2==0.9.6
python-dateutil==2.8.0
python3-openid==3.2.0
pytz==2019.3
pywin32-ctypes==0.2.0
pywinpty==0.5.5
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
Send2Trash==1.5.0
six==1.12.0
soupsieve==2.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.0
xlrd==1.2.0
yarl==1.4.2
yfinance==0.1.54
zipp==3.1.0
